Avoid internal server error when redis is not running.
When redis is not running express-session will not find a session store
ready and as a consequence not establish a req.session object. That led
to an error when running the integration test:

igelmac:connect2 dev$ DEBUG=express-session node_modules/.bin/mocha
--timeout 0  --compilers coffee:coffee-script/register
./test/integration/routes/pwless_signin_link.coffee  --reporter spec

  Passwordless signin link activation
      success flow existing user sign-IN
        express-session store is disconnected +0ms
        21:28:41 access: GET
        /signin/passwordless?token=token-id-random-stuff 500 23ms
...
  error:
     { [Error: cannot GET
     /signin/passwordless?token=token-id-random-stuff (500)]
          status: 500,
               text: 'Internal Server Error<br><br>Cannot set property
               \'amr\' of undefined',
                    method: 'GET',
                         path:
                         '/signin/passwordless?token=token-id-random-stuff'
                         },

This fix simply establishes an empty req.session object. It appears that
alternatives such as requiring to run redis for end to end integration
testing might be better.

Conflicts:
	test/integration/routes/pwless_signin_link.coffee