Remove mailer#getMailer method and remove caching
In oidc/requireVerifiedEmail.js the following pattern was used:

var mailer = require('../boot/mailer').getMailer()
 ... later in the code ... the mailer is referenced
 return function requireVerifiedEmail (req, res, next) {
     ...
     from: options.locals.from || mailer.from,

Having a mailer variable on the module level does not allow to easily swap in
test stubs or properties.

If this is done requireVerifiedEmail will continue to use the mailer as it was
initialized during the initial server setup.

The code above could have been changed to call getMailer()
in the requireVerifiedEmail(req, res, next) function body.

However it appeared simpler to simply remove the getMailer function
so that all mailer users would use the same module reference as cached
by node.